refactor(audio): split output.rs into render submodules
